Add domUpdate file with import error
